Convert all values to strings before storing them in the database
Unless the Database class does this manually, SQLite will attempt to do the serialization and will do it poorly. More specifically, it converts the number 1 to the string '1.0' and does not find any records when the column type is TEXT and the value provided to the query is a number.